+++ src/FScript.Runtime/ExternProvider.fs
@@ -1,0 +1,8 @@
+namespace FScript.Runtime
+
+open System
+
+[<AttributeUsage(AttributeTargets.Method, AllowMultiple = false)>]
+type FScriptExternProviderAttribute() =
+    inherit Attribute()
+